# perf: make `ape run --help` faster