--- a/src/ape_run/_cli.py
+++ b/src/ape_run/_cli.py
@@ -8,7 +8,6 @@ from runpy import run_module
 from typing import Any, Union
 
 import click
-from click import Command, Context, Option
 
 from ape.cli.commands import ConnectedProviderCommand
 from ape.cli.options import _VERBOSITY_VALUES, _create_verbosity_kwargs, verbosity_option
@@ -69,7 +68,7 @@ class ScriptCommand(click.MultiCommand):
         self._command_called = None
         self._has_warned_missing_hook: set[Path] = set()
 
-    def invoke(self, ctx: Context) -> Any:
+    def invoke(self, ctx: click.Context) -> Any:
         from ape.utils.basemodel import ManagerAccessMixin as access
 
         try:
@@ -96,10 +95,8 @@ class ScriptCommand(click.MultiCommand):
                 raise
 
     def _get_command(self, filepath: Path) -> Union[click.Command, click.Group, None]:
-        from ape.utils.basemodel import ManagerAccessMixin as access
-        from ape.utils.os import get_relative_path
-
-        relative_filepath = get_relative_path(filepath, access.local_project.path)
+        scripts_folder = Path.cwd() / "scripts"
+        relative_filepath = filepath.relative_to(scripts_folder)
 
         # First load the code module by compiling it
         # NOTE: This does not execute the module
@@ -126,14 +123,14 @@ class ScriptCommand(click.MultiCommand):
 
             self._namespace[filepath.stem] = cli_ns
             cli_obj = cli_ns["cli"]
-            if not isinstance(cli_obj, Command):
+            if not isinstance(cli_obj, click.Command):
                 logger.warning("Found `cli()` method but it is not a click command.")
                 return None
 
             params = [getattr(x, "name", None) for x in cli_obj.params]
             if "verbosity" not in params:
                 option_kwargs = _create_verbosity_kwargs()
-                option = Option(_VERBOSITY_VALUES, **option_kwargs)
+                option = click.Option(_VERBOSITY_VALUES, **option_kwargs)
                 cli_obj.params.append(option)
 
             cli_obj.name = filepath.stem if cli_obj.name in ("cli", "", None) else cli_obj.name
@@ -188,7 +185,7 @@ class ScriptCommand(click.MultiCommand):
         return self._get_cli_commands(scripts_folder)
 
     def _get_cli_commands(self, base_path: Path) -> dict:
-        commands: dict[str, Command] = {}
+        commands: dict[str, click.Command] = {}
 
         for filepath in base_path.iterdir():
             if filepath.stem.startswith("_"):
@@ -201,6 +198,7 @@ class ScriptCommand(click.MultiCommand):
                 subcommands = self._get_cli_commands(filepath)
                 for subcommand in subcommands.values():
                     group.add_command(subcommand)
+
                 commands[filepath.stem] = group
 
             if filepath.suffix == ".py":