### What I did

`ape run --help` was still slow because of a few reasons, mostly was the project access that i may look into on another PR.

### How I did it

Use `Path.cwd()` / "scripts"` locally.
Localize some imports as well.

Note: if you have a script module that is just slow to load, this wont help that ofc

### How to verify it

(in a project with zero scripts)

BEFORE:

```
ape run --help  4.47s user 0.99s system 98% cpu 5.542 total
```

AFTER:

```
ape run --help  0.29s user 0.09s system 59% cpu 0.638 total
```

With scripts is also much faster, but you wil only be as fast as your slowest script to load in a project. CLI users should have the same knowledge that lazy imports are better for CLIs.

### Checklist

<!-- All PRs must complete the following checklist before being merged -->

- [ ] All changes are completed
- [ ] New test cases have been added
- [ ] Documentation has been updated